# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StylistTG API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs/profile/preview": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Preview Profile Job",
        "operationId": "preview_profile_job_api_jobs_profile_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilePreviewRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/profile": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Post Profile Job",
        "operationId": "post_profile_job_api_jobs_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummaryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Endpoint",
        "operationId": "get_job_endpoint_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Delete Job Endpoint",
        "operationId": "delete_job_endpoint_api_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/steps": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Steps Endpoint",
        "operationId": "get_job_steps_endpoint_api_jobs__job_id__steps_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobStepListItemRead"
                  },
                  "title": "Response Get Job Steps Endpoint Api Jobs  Job Id  Steps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job Endpoint",
        "operationId": "cancel_job_endpoint_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummaryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobDetailRead": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "job_state": {
            "type": "string",
            "title": "Job State"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "execution_intent_hash": {
            "type": "string",
            "title": "Execution Intent Hash"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "can_retry": {
            "type": "boolean",
            "title": "Can Retry"
          },
          "can_refresh_runtime": {
            "type": "boolean",
            "title": "Can Refresh Runtime"
          },
          "step_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Step Counts"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "job_state",
          "account_id",
          "execution_intent_hash",
          "started_at",
          "finished_at",
          "failure_reason",
          "can_retry",
          "can_refresh_runtime",
          "step_counts"
        ],
        "title": "JobDetailRead"
      },
      "JobStepListItemRead": {
        "properties": {
          "step_key": {
            "type": "string",
            "title": "Step Key"
          },
          "step_type": {
            "type": "string",
            "title": "Step Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verification_attempted": {
            "type": "boolean",
            "title": "Verification Attempted"
          },
          "verification_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Result"
          },
          "uncertain_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uncertain Reason"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Class"
          },
          "result_payload_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Payload Json"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "step_key",
          "step_type",
          "status",
          "verification_attempted",
          "verification_result",
          "uncertain_reason",
          "error_code",
          "error_class",
          "result_payload_json",
          "started_at",
          "finished_at"
        ],
        "title": "JobStepListItemRead"
      },
      "JobSummaryRead": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "job_state": {
            "type": "string",
            "title": "Job State"
          },
          "execution_intent_hash": {
            "type": "string",
            "title": "Execution Intent Hash"
          },
          "plan_summary": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Plan Summary"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "dedup_blocked_by_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedup Blocked By Job Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "job_state",
          "execution_intent_hash",
          "plan_summary",
          "created_at"
        ],
        "title": "JobSummaryRead"
      },
...
      "ProfileJobCreate": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "photo_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Asset Id"
          },
          "mock_fail_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mock Fail Step"
          },
          "mock_crash_after_step_started": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mock Crash After Step Started"
          }
        },
        "type": "object",
        "required": [
          "account_id"
        ],
        "title": "ProfileJobCreate"
      },
      "ProfilePreviewRead": {
        "properties": {
          "can_create_job": {
            "type": "boolean",
            "title": "Can Create Job"
          },
          "blocking_errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocking Errors"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "normalized_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Normalized Payload"
          },
          "execution_intent_hash": {
            "type": "string",
            "title": "Execution Intent Hash"
          },
          "plan_json_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Plan Json Snapshot"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ProfilePreviewStepRead"
            },
            "type": "array",
            "title": "Steps"
          },
          "requires_execution_usable": {
            "type": "boolean",
            "title": "Requires Execution Usable"
          },
          "dedup_would_block": {
            "type": "boolean",
            "title": "Dedup Would Block"
          },
          "dedup_blocked_by_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedup Blocked By Job Id"
          }
        },
        "type": "object",
        "required": [
          "can_create_job",
          "blocking_errors",
          "warnings",
          "normalized_payload",
          "execution_intent_hash",
          "plan_json_snapshot",
          "steps",
          "requires_execution_usable",
          "dedup_would_block",
          "dedup_blocked_by_job_id"
        ],
        "title": "ProfilePreviewRead"
      },
      "ProfilePreviewRequest": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "photo_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Asset Id"
          }
        },
        "type": "object",
        "required": [
          "account_id"
        ],
        "title": "ProfilePreviewRequest"
      },
      "ProfilePreviewStepRead": {
        "properties": {
          "step_key": {
            "type": "string",
            "title": "Step Key"
          },
          "step_type": {
            "type": "string",
            "title": "Step Type"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "idempotency_class": {
            "type": "string",
            "title": "Idempotency Class"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "step_key",
          "step_type",
          "order",
          "required",
          "idempotency_class",
          "payload"
        ],
        "title": "ProfilePreviewStepRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}